Refactoring and normalizing as agreed in PR - 1
