+++ src/Components/Item/styles.js
@@ -1,0 +1,8 @@
+export const styles = {
+  card: {
+    margin: 10
+  },
+  media: {
+    height: 200
+  }
+};
