--- src/Components/Layout/index.js
@@ -1,3 +1,0 @@
-import Header from "./Header";
-
-export { Header };
