+++ src/Components/Layout/styles.js
@@ -1,0 +1,12 @@
+export const styles = {
+  root: {
+    flexGrow: 1
+  },
+  flex: {
+    flex: 1
+  },
+  menuButton: {
+    marginLeft: -12,
+    marginRight: 20
+  }
+};
