--- src/Components/ListItems/index.js
@@ -1,3 +1,0 @@
-import ListItems from "./List";
-
-export { ListItems };
