+++ src/Components/SearchBar/styles.js
@@ -1,0 +1,13 @@
+export const styles = {
+  container: {
+    padding: "10px 50px",
+    marginTop: 15
+  },
+  Paper: {
+    padding: 20
+  },
+  Button: {
+    marginTop: "10%",
+    marginLeft: "5%"
+  }
+};
